+++ pkg/containers/tracker.go
@@ -67,37 +67,6 @@ type StepTagger interface {
 	OrdinalAt(t time.Time) uint32
 }
 
-// Enricher is a late-binding events.ContainerEnricher. The event pipeline
-// starts early in daemon startup, but container tracking can only start
-// after the dockerd restart severs-proofs the event subscription — so the
-// pipeline gets this shell up front and Bind attaches the live Tracker once
-// it exists. Until then (and if tracking never comes up) Enrich is a no-op,
-// which is the correct degradation: unenriched events classify to the
-// stricter tiers.
-type Enricher struct {
-	tracker atomic.Pointer[Tracker]
-}
-
-// Bind attaches the live tracker; safe to call once tracking starts.
-func (e *Enricher) Bind(t *Tracker) { e.tracker.Store(t) }
-
-// Enrich implements events.ContainerEnricher.
-func (e *Enricher) Enrich(audit *events.AuditEvent, ev *events.BpfBlockedEvent) {
-	if t := e.tracker.Load(); t != nil {
-		t.Enrich(audit, ev)
-	}
-}
-
-// DecorateVerdict adds container identity to a cgroup-hook outcome when the
-// tracker is live. A verdict that beats docker tracking startup simply goes
-// undecorated — it is still reported in full by pkg/events, just without a
-// container id.
-func (e *Enricher) DecorateVerdict(audit *events.AuditEvent, rec origin.Record) {
-	if t := e.tracker.Load(); t != nil {
-		t.DecorateVerdict(audit, rec)
-	}
-}
-
 // Options configures Start. Zero values take the production defaults; the
 // roots are injectable for tests.
 type Options struct {
@@ -395,719 +364,6 @@ func (t *Tracker) streamOnce(ctx context.Context) (healthy bool, err error) {
 			t.remove(ev.Actor.ID)
 		}
 	}
-}
-
-// reconcile treats every running container we don't know as a start event of
-// kind "reconcile". Containers that predate the tracker (job containers and
-// services started before cargowall, or anything running across a
-// reconnect) get StepOrdinalPreDaemon — mirroring seedExisting's semantics
-// for pre-daemon processes — and their exec events re-tag real step
-// ordinals from there.
-//
-// It also runs the diff the OTHER way: a tracked container absent from the
-// list lost its die/destroy somewhere the `since` replay cannot reach (a
-// dockerd restart empties the daemon's event buffer). Without the removal
-// its stale byIP entry keeps answering DNS attribution lookups, crediting
-// whatever later occupies that bridge address to a dead container.
-func (t *Tracker) reconcile(ctx context.Context) {
-	list, err := t.client.listContainers(ctx)
-	if err != nil {
-		t.logger.Warn("Cannot list containers for reconciliation", "error", err)
-		return
-	}
-	live := make(map[string]bool, len(list))
-	for _, c := range list {
-		live[c.ID] = true
-	}
-	// Split the list: unknown IDs are adopted (below), known IDs are
-	// verified — a known ID may be a NEW incarnation, because docker
-	// restart keeps the container ID but mints a fresh init PID and cgroup,
-	// and the die+start pair can both be lost in the same stream gap.
-	// Verify by init PID; on mismatch re-adopt, or byCgroup keys a dead
-	// inode and the new init process is never tagged.
-	type incarnation struct {
-		id             string
-		oldPID, newPID int
-	}
-	var toVerify []incarnation
-	for _, c := range list {
-		t.mu.Lock()
-		info := t.containers[c.ID]
-		t.mu.Unlock()
-		if info == nil {
-			t.handleStart(ctx, c.ID, "reconcile", 0)
-			continue
-		}
-		toVerify = append(toVerify, incarnation{id: c.ID, oldPID: info.initPID})
-	}
-	// Verification inspects run concurrently, bounded: this whole function
-	// runs on the stream goroutine before event processing resumes, and a
-	// large fleet against a slow daemon must not serialize N unary
-	// timeouts in front of reconnect recovery.
-	var (
-		wg      sync.WaitGroup
-		sem     = make(chan struct{}, reconcileInspectConcurrency)
-		mu      sync.Mutex
-		readopt []incarnation
-	)
-	for _, k := range toVerify {
-		wg.Add(1)
-		go func(k incarnation) {
-			defer wg.Done()
-			select {
-			case sem <- struct{}{}:
-				defer func() { <-sem }()
-			case <-ctx.Done():
-				return
-			}
-			ictx, cancel := context.WithTimeout(ctx, unaryTimeout)
-			insp, ierr := t.client.inspectContainer(ictx, k.id)
-			cancel()
-			if ierr != nil || insp.State.Pid == 0 || insp.State.Pid == k.oldPID {
-				return
-			}
-			k.newPID = insp.State.Pid
-			mu.Lock()
-			readopt = append(readopt, k)
-			mu.Unlock()
-		}(k)
-	}
-	wg.Wait()
-	for _, k := range readopt {
-		t.logger.Info("Container re-adopted during reconciliation (restarted across a stream gap)",
-			"container", shortID(k.id), "old_pid", k.oldPID, "new_pid", k.newPID)
-		t.remove(k.id)
-		t.handleStart(ctx, k.id, "reconcile", 0)
-	}
-
-	// The diff the other way: a tracked container absent from the list lost
-	// its die/destroy somewhere the `since` replay cannot reach. A container
-	// started after the list call is either absent from t.containers (its
-	// start event is still buffered in the stream — the sweep leaves it
-	// alone) or was added by a handleStart above (in the list, so live).
-	// Only genuinely dead containers are swept — unless the list came back
-	// EMPTY while we track containers, which on a just-restarted dockerd is
-	// far more plausibly an initializing daemon than a mass die-off; wiping
-	// the indexes on that evidence would permanently demote every live
-	// container (live-restore containers never re-emit a start event).
-	// Absence must be seen on TWO consecutive passes before the sweep
-	// fires: a settling daemon can answer /containers/json with a partial
-	// list, and a single partial answer must not wipe live containers
-	// (live-restore containers never re-emit a start event, so a wrong
-	// sweep is permanent). The empty-list guard below is the degenerate
-	// case of the same distrust.
-	t.mu.Lock()
-	var stale []string
-	for id := range t.containers {
-		if live[id] {
-			delete(t.missingCounts, id)
-			continue
-		}
-		t.missingCounts[id]++
-		if t.missingCounts[id] >= 2 {
-			stale = append(stale, id)
-			delete(t.missingCounts, id)
-		}
-	}
-	tracked := len(t.containers)
-	t.mu.Unlock()
-	if len(list) == 0 && tracked > 0 {
-		t.logger.Warn("Reconciliation list empty while containers are tracked; keeping indexes (initializing daemon?)",
-			"tracked", tracked)
-		return
-	}
-	for _, id := range stale {
-		t.logger.Info("Container removed during reconciliation (absent from two consecutive lists)",
-			"container", shortID(id))
-		t.remove(id)
-	}
-
-	// seenExecs sets can outlive their container: an exec handler parked in
-	// resolveExecPid may re-create one after remove() ran (a race
-	// markExecSeen deliberately tolerates). The container sweep above walks
-	// t.containers, which such an orphan is not in — prune it here.
-	t.mu.Lock()
-	for id := range t.seenExecs {
-		if !live[id] && t.containers[id] == nil {
-			delete(t.seenExecs, id)
-		}
-	}
-	t.mu.Unlock()
-}
-
-func (t *Tracker) handleStart(ctx context.Context, id, kind string, timeNano int64) {
-	t.mu.Lock()
-	existing := t.containers[id]
-	t.mu.Unlock()
-	if existing != nil {
-		// Replayed event after reconnect — usually a no-op, EXCEPT when
-		// reconcile got there first and adopted the container as pre-daemon:
-		// the replayed start carries the real timeNano the sentinel adoption
-		// lacked, and discarding it would leave the container filed under
-		// "started before cargowall" for the run. Upgrade instead.
-		if kind == "start" && timeNano > 0 && existing.preDaemon {
-			t.upgradeReconciledStart(existing, id, timeNano)
-		}
-		return
-	}
-
-	ictx, cancel := context.WithTimeout(ctx, unaryTimeout)
-	insp, err := t.client.inspectContainer(ictx, id)
-	cancel()
-	if err != nil {
-		t.logger.Debug("Container inspect failed", "container", shortID(id), "error", err)
-		return
-	}
-	if insp.State.Pid == 0 {
-		return // already gone; die/destroy would have nothing to remove
-	}
-
-	info := &containerInfo{
-		id:         insp.ID,
-		initPID:    insp.State.Pid,
-		privileged: insp.HostConfig.Privileged,
-	}
-	for _, ipStr := range collectIPs(insp) {
-		if ip, err := netip.ParseAddr(ipStr); err == nil {
-			info.ips = append(info.ips, ip)
-		}
-	}
-
-	// New bridge subnets are carved out before any tagging concerns: the
-	// network exists whatever the identity check below decides.
-	if t.opts.AllowLocalSubnet != nil {
-		for _, s := range collectSubnets(insp) {
-			t.allowSubnetOnce(ctx, s)
-		}
-	}
-
-	// Identity check before any tagging: the inspected PID must currently
-	// sit in this container's cgroup, or a recycled PID could tag an
-	// unrelated host process.
-	cgroupID, err := verifyContainerTask(t.opts.ProcRoot, t.opts.CgroupRoot, info.initPID, insp.ID)
-	identityOK := err == nil
-	if !identityOK {
-		t.identityRejected.Add(1)
-		t.logger.Warn("Container identity check failed; not tagging", "container", shortID(id), "error", err)
-	}
-	info.cgroupID = cgroupID
-
-	var ordinal uint32
-	if kind == "reconcile" {
-		ordinal = events.StepOrdinalPreDaemon
-		info.preDaemon = true
-	} else {
-		ordinal = t.tagger.OrdinalAt(time.Unix(0, timeNano))
-	}
-	// The IP indexes keep the event-time ordinal even when tagging below is
-	// refused: DNS attribution keys on the container's address (from
-	// inspect), which the identity check — a guard against tagging a
-	// recycled PID — says nothing about.
-	info.birthOrdinal = ordinal
-	info.effectiveOrdinal = ordinal
-
-	// markerOrdinal is what the container_attribution audit event reports:
-	// the TAGGING outcome. A refused or skipped tag reports 0 so consumers
-	// (and the CI assertions) can never mistake "would have been step N"
-	// for "tagged with step N".
-	markerOrdinal := uint32(0)
-	var latencyMS float64
-	if identityOK && ordinal != 0 {
-		t.tagger.TagContainerProcess(info.initPID, ordinal)
-		markerOrdinal = ordinal
-		if kind == "reconcile" {
-			t.reconciled.Add(1)
-		} else {
-			t.tagged.Add(1)
-			latencyMS = float64(time.Since(time.Unix(0, timeNano)).Microseconds()) / 1000
-			t.recordLatency(latencyMS)
-		}
-	} else if ordinal == 0 {
-		// Started before any step boundary (or outside a workflow): stays
-		// untagged and classifies to the container-unattributed tier.
-		t.tagSkipped.Add(1)
-	}
-
-	if info.privileged {
-		t.logger.Warn("Privileged container started (host-root equivalent; observer coverage not guaranteed)",
-			"container", shortID(id))
-	}
-
-	t.mu.Lock()
-	// Newest inspect wins a conflicted address: for docker-IPAM recycling
-	// (die/destroy lost in a stream gap) the newcomer genuinely holds the
-	// address, and for manually-assigned duplicates (macvlan/static --ip,
-	// where IPAM guarantees nothing) this matches the pre-existing
-	// last-writer semantics. The displaced holder is NOT untracked — its
-	// cgroup identity and other addresses may be live; if it is truly
-	// dead, reconcile's sweep collects it. Logged after unlock: both
-	// ringbuf reader paths contend on t.mu, and a slow slog handler must
-	// not stall them.
-	var displaced []string
-	for _, ip := range info.ips {
-		if prev := t.byIP[ip]; prev != nil && prev != info {
-			displaced = append(displaced, shortID(prev.id)+" lost "+ip.String())
-		}
-	}
-	t.containers[insp.ID] = info
-	for _, ip := range info.ips {
-		t.byIP[ip] = info
-	}
-	if info.cgroupID != 0 {
-		t.byCgroup[info.cgroupID] = info
-	}
-	t.mu.Unlock()
-
-	for _, d := range displaced {
-		t.logger.Info("Address conflict: newest inspect wins", "displaced", d, "winner", shortID(id))
-	}
-
-	t.logger.Info("Container attributed",
-		"container", shortID(id), "kind", kind, "pid", info.initPID,
-		"step_ordinal", markerOrdinal, "tag_latency_ms", latencyMS, "privileged", info.privileged)
-	t.logAttribution(events.AuditEvent{
-		EventType:       events.EventContainerAttribution,
-		ContainerID:     shortID(id),
-		ContainerOrigin: true,
-		AttributionKind: kind,
-		StepOrdinal:     markerOrdinal,
-		PID:             uint32(info.initPID),
-		TagLatencyMS:    latencyMS,
-		Privileged:      info.privileged,
-	})
-}
-
-// upgradeReconciledStart replaces a pre-daemon sentinel adoption with the
-// real attribution its replayed start event carries: reconcile runs before
-// the since-replay is drained, so a container born in a stream gap is
-// adopted as StepOrdinalPreDaemon moments before its true start event —
-// with the correct event-time ordinal — arrives. Without the upgrade the
-// replay is discarded and the container reports as "started before
-// cargowall" for the rest of the run.
-func (t *Tracker) upgradeReconciledStart(info *containerInfo, id string, timeNano int64) {
-	ordinal := t.tagger.OrdinalAt(time.Unix(0, timeNano))
-	t.mu.Lock()
-	info.preDaemon = false
-	if ordinal == 0 {
-		t.mu.Unlock()
-		return // genuinely pre-step: the sentinel was right
-	}
-	info.birthOrdinal = ordinal
-	if info.effectiveOrdinal == events.StepOrdinalPreDaemon {
-		// Keep a real exec re-tag if one already happened.
-		info.effectiveOrdinal = ordinal
-	}
-	pid := info.initPID
-	t.mu.Unlock()
-
-	// Identity was verified at adoption; the PID has not changed since.
-	t.tagger.TagContainerProcess(pid, ordinal)
-	t.tagged.Add(1)
-	t.logger.Info("Container attribution upgraded from replayed start event",
-		"container", shortID(id), "step_ordinal", ordinal)
-	t.logAttribution(events.AuditEvent{
-		EventType:       events.EventContainerAttribution,
-		ContainerID:     shortID(id),
-		ContainerOrigin: true,
-		AttributionKind: "start",
-		StepOrdinal:     ordinal,
-		PID:             uint32(pid),
-	})
-}
-
-func (t *Tracker) handleExecStart(ctx context.Context, containerID, execID string, timeNano int64) {
-	pid, err := t.resolveExecPid(ctx, containerID, execID)
-	if err != nil {
-		if errors.Is(err, errExecAlreadySeen) {
-			// Replayed exec_start after a stream reconnect (docker's `since`
-			// is inclusive): handled once already. A second pass would
-			// re-inspect, double-count execTagged, skew the latency
-			// percentiles with an event minutes old, and re-tag a PID that
-			// may since have been recycled.
-			return
-		}
-		t.tagSkipped.Add(1)
-		t.logger.Debug("Exec PID unresolvable; traffic stays in the container tier",
-			"container", shortID(containerID), "error", err)
-		return
-	}
-
-	// Exec-time ordinal, never the container's birth ordinal: a `docker
-	// exec` from step B into a container step A created belongs to B.
-	ordinal := t.tagger.OrdinalAt(time.Unix(0, timeNano))
-	if ordinal == 0 {
-		t.tagSkipped.Add(1)
-		return
-	}
-
-	if _, err := verifyContainerTask(t.opts.ProcRoot, t.opts.CgroupRoot, pid, containerID); err != nil {
-		t.identityRejected.Add(1)
-		t.logger.Warn("Exec identity check failed; not tagging", "container", shortID(containerID), "error", err)
-		return
-	}
-
-	t.tagger.TagContainerProcess(pid, ordinal)
-	t.execTagged.Add(1)
-	latencyMS := float64(time.Since(time.Unix(0, timeNano)).Microseconds()) / 1000
-	t.recordLatency(latencyMS)
-
-	t.mu.Lock()
-	if info := t.containers[containerID]; info != nil {
-		info.effectiveOrdinal = ordinal
-	}
-	t.mu.Unlock()
-
-	t.logger.Info("Container exec attributed",
-		"container", shortID(containerID), "pid", pid,
-		"step_ordinal", ordinal, "tag_latency_ms", latencyMS)
-	t.logAttribution(events.AuditEvent{
-		EventType:       events.EventContainerAttribution,
-		ContainerID:     shortID(containerID),
-		ContainerOrigin: true,
-		AttributionKind: "exec",
-		StepOrdinal:     ordinal,
-		PID:             uint32(pid),
-		TagLatencyMS:    latencyMS,
-	})
-}
-
-// errExecAlreadySeen means the exec was handled on a previous delivery — a
-// reconnect replay, not a failure; handleExecStart returns without counting
-// or logging it as a skip.
-var errExecAlreadySeen = errors.New("exec already handled")
-
-// resolveExecPid turns an exec event into the exec leader's host PID. The
-// execID attribute is the fast path; daemons that omit it fall back to
-// diffing the container's live ExecIDs against what we've already handled.
-// Both paths honour the seenExecs idempotency guard — the replay after a
-// stream reconnect arrives on whichever path the daemon supports.
-// The error exists for the Debug log: an unresolvable exec is expected
-// degradation (exec already exited), but WHICH failure occurred must be
-// diagnosable from a debug run.
-func (t *Tracker) resolveExecPid(ctx context.Context, containerID, execID string) (int, error) {
-	ictx, cancel := context.WithTimeout(ctx, unaryTimeout)
-	defer cancel()
-
-	if execID != "" {
-		if t.markExecSeen(containerID, execID) {
-			return 0, errExecAlreadySeen
-		}
-		return t.inspectExecPid(ictx, execID)
-	}
-
-	cinsp, err := t.client.inspectContainer(ictx, containerID)
-	if err != nil {
-		return 0, fmt.Errorf("container inspect for exec fallback: %w", err)
-	}
-	sawSeen := false
-	for _, id := range cinsp.ExecIDs {
-		if t.markExecSeen(containerID, id) {
-			sawSeen = true
-			continue
-		}
-		if pid, err := t.inspectExecPid(ictx, id); err == nil {
-			return pid, nil
-		}
-	}
-	if sawSeen {
-		// Every live exec was already handled — the signature of a replayed
-		// event after a stream reconnect, so the guard must hold on this
-		// path exactly as on the execID fast path. (Approximation: an
-		// unseen exec that failed inspection alongside a seen one is also
-		// classified as replay; it would have been unattributable anyway.)
-		return 0, errExecAlreadySeen
-	}
-	return 0, fmt.Errorf("no execID attribute and no unseen running exec among %d", len(cinsp.ExecIDs))
-}
-
-// markExecSeen records execID under its container and reports whether it
-// had already been handled. Per-container keying exists so remove() can
-// reclaim the whole set when the container goes away.
-func (t *Tracker) markExecSeen(containerID, execID string) (seen bool) {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-	set := t.seenExecs[containerID]
-	if set == nil {
-		set = make(map[string]bool)
-		t.seenExecs[containerID] = set
-	}
-	seen = set[execID]
-	set[execID] = true
-	return seen
-}
-
-// inspectExecPid reads an exec's leader PID, retrying briefly: dockerd
-// emits exec_start BEFORE it records the process pid from containerd, so an
-// immediate inspect reads Pid 0 for an exec that is very much alive
-// (confirmed against dockerd 28/29; same publish-race shape as the
-// stepCmdline retry in pkg/steps). A genuinely exited exec keeps reading 0
-// and falls out as unattributable — the stricter tier, by design. The retry
-// window is small against any real exec workload and is included in the
-// reported tag latency.
-func (t *Tracker) inspectExecPid(ctx context.Context, execID string) (int, error) {
-	for i := 0; ; i++ {
-		insp, err := t.client.inspectExec(ctx, execID)
-		if err != nil {
-			return 0, fmt.Errorf("exec inspect %s: %w", shortID(execID), err)
-		}
-		if insp.Pid != 0 {
-			return insp.Pid, nil
-		}
-		if i >= 19 {
-			return 0, fmt.Errorf("exec %s has no pid after retries (already exited?)", shortID(execID))
-		}
-		select {
-		case <-ctx.Done():
-			return 0, fmt.Errorf("exec %s pid wait: %w", shortID(execID), ctx.Err())
-		case <-time.After(50 * time.Millisecond):
-		}
-	}
-}
-
-func (t *Tracker) remove(id string) {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-	// Unconditional: exec ids may have been seen for containers the start
-	// handler never registered (inspect raced the container's death).
-	delete(t.seenExecs, id)
-	info := t.containers[id]
-	if info == nil {
-		return
-	}
-	delete(t.containers, id)
-	for _, ip := range info.ips {
-		if t.byIP[ip] == info {
-			delete(t.byIP, ip)
-		}
-	}
-	if info.cgroupID != 0 && t.byCgroup[info.cgroupID] == info {
-		delete(t.byCgroup, info.cgroupID)
-	}
-	// Task-map cleanup is the kernel's sched_process_exit; nothing to undo.
-}
-
-// forEachInspectAddr visits every address docker reported for a container:
-// the legacy top-level fields (the default bridge; network name "") and
-// each named network's entry. THE one walker of the NetworkSettings shape —
-// both DNS attribution (collectIPs) and the enforce-mode subnet carve-out
-// (collectSubnets) ride it, so a new address source added here reaches
-// both, and a source added elsewhere is a bug.
-func forEachInspectAddr(insp containerInspect, visit func(network, ip string, prefixLen int)) {
-	visit("", insp.NetworkSettings.IPAddress, insp.NetworkSettings.IPPrefixLen)
-	visit("", insp.NetworkSettings.GlobalIPv6Address, insp.NetworkSettings.GlobalIPv6PrefixLen)
-	for name, nw := range insp.NetworkSettings.Networks {
-		visit(name, nw.IPAddress, nw.IPPrefixLen)
-		visit(name, nw.GlobalIPv6Address, nw.GlobalIPv6PrefixLen)
-	}
-}
-
-// collectIPs gathers every address docker assigned the container, v4 and
-// v6 alike: the DNS client lookup keys on whichever family the container's
-// resolver socket uses, so an IPv6-enabled bridge must not silently lose
-// DNS attribution while the TC join (cgroup-id based) keeps working.
-func collectIPs(insp containerInspect) []string {
-	ips := []string{}
-	forEachInspectAddr(insp, func(_, ip string, _ int) {
-		if ip != "" && !slices.Contains(ips, ip) {
-			ips = append(ips, ip)
-		}
-	})
-	return ips
-}
-
-// localSubnet is one candidate carve-out: a subnet plus the network it was
-// seen on ("" = the default bridge, which is bridge-driver by definition).
-type localSubnet struct {
-	network string
-	prefix  netip.Prefix
-}
-
-// collectSubnets derives the subnets docker attached the container to, v4
-// and v6, masked. Entries with no prefix length (host networking, daemons
-// that omit it) are dropped — a zero prefix would shape into a catch-all.
-// Driver gating happens later, in allowSubnetOnce: this function only
-// reads the inspect payload.
-func collectSubnets(insp containerInspect) []localSubnet {
-	var subnets []localSubnet
-	forEachInspectAddr(insp, func(network, ipStr string, prefixLen int) {
-		if ipStr == "" || prefixLen <= 0 {
-			return
-		}
-		addr, err := netip.ParseAddr(ipStr)
-		if err != nil {
-			return
-		}
-		prefix, err := addr.Prefix(prefixLen)
-		if err != nil {
-			return // prefixLen wider than the address family
-		}
-		for _, s := range subnets {
-			if s.prefix == prefix {
-				return
-			}
-		}
-		subnets = append(subnets, localSubnet{network: network, prefix: prefix})
-	})
-	return subnets
-}
-
-// allowSubnetOnce runs the once-per-subnet carve-out pipeline: resolve the
-// network's driver (bridge-driver networks are local-only by construction —
-// their routes are on-link; a macvlan/ipvlan subnet is physical-network
-// space and MUST stay adjudicated), then hand the subnet to the callback.
-// seenSubnets is stamped only once the decision is final — a failed driver
-// inspect or a failed carve-out write is retried by the next container on
-// the network.
-func (t *Tracker) allowSubnetOnce(ctx context.Context, s localSubnet) {
-	key := s.prefix.String()
-	t.mu.Lock()
-	seen := t.seenSubnets[key]
-	t.mu.Unlock()
-	if seen {
-		return
-	}
-	// The legacy top-level inspect fields (network "") are resolved to the
-	// literal "bridge" network and driver-checked like any other: stock
-	// moby only populates them from the default bridge, but nothing
-	// guarantees that across docker-compatible daemons, and the macvlan
-	// protection must not rest on undocumented daemon behavior.
-	netName := s.network
-	if netName == "" {
-		netName = "bridge"
-	}
-	t.mu.Lock()
-	isBridge, known := t.netDrivers[netName]
-	t.mu.Unlock()
-	if !known {
-		ictx, cancel := context.WithTimeout(ctx, unaryTimeout)
-		nw, err := t.client.inspectNetwork(ictx, netName)
-		cancel()
-		if err != nil {
-			t.logger.Warn("Cannot inspect network for subnet carve-out; subnet stays adjudicated for now",
-				"network", netName, "subnet", key, "error", err)
-			return // unresolved: retried by the next container, not cached
-		}
-		isBridge = nw.Driver == "bridge"
-		t.mu.Lock()
-		t.netDrivers[netName] = isBridge
-		t.mu.Unlock()
-	}
-	if !isBridge {
-		t.mu.Lock()
-		t.seenSubnets[key] = true
-		t.mu.Unlock()
-		t.logger.Info("Non-bridge network subnet stays adjudicated (not local-only)",
-			"network", netName, "subnet", key)
-		return
-	}
-	if err := t.opts.AllowLocalSubnet(s.prefix); err != nil {
-		t.logger.Warn("Local subnet carve-out failed; will retry on the network's next container",
-			"subnet", key, "error", err)
-		return
-	}
-	t.mu.Lock()
-	t.seenSubnets[key] = true
-	t.mu.Unlock()
-}
-
-// handleNetworkCreate carves a just-created bridge network's subnets
-// BEFORE its first container can produce traffic — container-driven
-// discovery alone misses networks whose containers exit faster than an
-// inspect (docker run --rm ... true), and under enforce an uncarved bridge
-// is a broken network.
-func (t *Tracker) handleNetworkCreate(ctx context.Context, ev dockerEvent) {
-	if t.opts.AllowLocalSubnet == nil {
-		return
-	}
-	name := ev.Actor.Attributes["name"]
-	if name == "" {
-		name = ev.Actor.ID
-	}
-	ictx, cancel := context.WithTimeout(ctx, unaryTimeout)
-	nw, err := t.client.inspectNetwork(ictx, name)
-	cancel()
-	if err != nil {
-		t.logger.Warn("Cannot inspect created network; its subnets stay adjudicated until a container appears",
-			"network", name, "error", err)
-		return
-	}
-	isBridge := nw.Driver == "bridge"
-	t.mu.Lock()
-	t.netDrivers[name] = isBridge
-	t.mu.Unlock()
-	if !isBridge {
-		return
-	}
-	for _, cfg := range nw.IPAM.Config {
-		prefix, perr := netip.ParsePrefix(cfg.Subnet)
-		if perr != nil {
-			continue
-		}
-		prefix = prefix.Masked()
-		key := prefix.String()
-		t.mu.Lock()
-		seen := t.seenSubnets[key]
-		t.mu.Unlock()
-		if seen {
-			continue
-		}
-		if aerr := t.opts.AllowLocalSubnet(prefix); aerr != nil {
-			t.logger.Warn("Local subnet carve-out failed; will retry via container discovery",
-				"subnet", key, "error", aerr)
-			continue
-		}
-		t.mu.Lock()
-		t.seenSubnets[key] = true
-		t.mu.Unlock()
-	}
-}
-
-// DiscoverBridgeSubnets returns every bridge-driver network's IPAM
-// subnets. cmd/start.go runs this BEFORE raising the cgroup hook out of
-// observe: docker-event tracking starts only after the dockerd restart, and
-// networks that predate cargowall — including ones whose containers are
-// stopped, or are stopped by that very restart — must not lose bridge-local
-// traffic in the window before tracking begins. Enumerating networks
-// rather than containers means one API call, no per-container inspects on
-// the timing-sensitive startup path, and no dependency on anything being
-// alive to inspect.
-func DiscoverBridgeSubnets(ctx context.Context, socket string) ([]netip.Prefix, error) {
-	if socket == "" {
-		socket = "/var/run/docker.sock"
-	}
-	c := newDockerClient(socket)
-	pingCtx, cancel := context.WithTimeout(ctx, 2*time.Second)
-	err := c.ping(pingCtx)
-	cancel()
-	if err != nil {
-		return nil, fmt.Errorf("docker daemon unreachable: %w", err)
-	}
-	lctx, cancel := context.WithTimeout(ctx, unaryTimeout)
-	nets, err := c.listNetworks(lctx)
-	cancel()
-	if err != nil {
-		return nil, fmt.Errorf("list networks: %w", err)
-	}
-	seen := map[netip.Prefix]bool{}
-	var out []netip.Prefix
-	for _, nw := range nets {
-		if nw.Driver != "bridge" {
-			continue
-		}
-		for _, cfg := range nw.IPAM.Config {
-			prefix, perr := netip.ParsePrefix(cfg.Subnet)
-			if perr != nil {
-				continue
-			}
-			prefix = prefix.Masked()
-			if seen[prefix] {
-				continue
-			}
-			seen[prefix] = true
-			out = append(out, prefix)
-		}
-	}
-	return out, nil
 }
 
 func (t *Tracker) logAttribution(ev events.AuditEvent) {
